backend/requirements.txt
torch==2.2.0
torchvision==0.17.0
dalle-pytorch==1.6.6
pillow==10.3.0
tqdm==4.66.4
rich==13.7.0             # For beautiful CLI output
